requirements.txt
flask
flask-session
flask-socketio
google-auth